backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
openai
python-dotenv
httpx
